Create Task1(A).cpp
Step 1 :    Store the price of toffees in vector.
Step 2 :    Using sort function sort the vector in ascending order.
Step 3 :    Iterate through the vector.
Step 4 :    If ith element of vector is less than money , to buy the toffee subtract the value of ith element from money. 
                 And keep a count of the tofee bought.
Step 5 :    Else  break because the vector is sorted so the next element will also be greater than money.